requirements.txt
pycryptodome==3.15.0
pydantic[dotenv]==1.9.1
Pyrogram==2.0.30
TgCrypto==1.2.3
